git everything working nice n frosty
