--- .secret.yml.example
@@ -1,1 +1,0 @@
-OPENAI_API_KEY: sk-N0y...
